elf: consider cases when program segments are located at the end of the file
Program segments can be placed at the end of the file, up till now we did not
calculate the program segment end hence ELF files where program segments
were placed at the end of the file were miss calculated. This could happen
when there are no sections in the ELF file, which is unusual though possible.