Improve cramfs CRC checking.
We observed a sample from Cisco where the CRC value was stored in a
different endianness. This is non standard but we decided to support it
anyway.

We also added a check for old cramfs format. If it's an old format, the
CRC check always returns true because it's not supported.

More information about CRC validation and old format:
https://github.com/secularbird/cramfs/blob/master/cramfsck.c