Remove magic_file API
We introduced this API to be able to use it ourselves, but as it turns out we
don't need it, and we don't want to expose a public API unnecessarily, because
we can never "take it back". Also the `MAGIC` environment variable works well for
libmagic if someone wants to control the magic database, so this API might not
even be needed.

The problem we were trying to solve to use the same magic database on different
systems, but that turned out impossible to do, because the magic database format
breaks between every version, so the only way to use libmagic is to use the same
version of the magic database, which is usually the one shipped with the system.