TEMP: zig build chain instead of Rust
The whole thing can be part of dev-dependencies, fully automated, the user
don't have to download a Rust toolchain and the compiled shared library
can be included in the package.

There is setuptools-zig package, but it cannot be used with the "ziglang"
package, so I just copied it's code, modified it to make it work.