Correct formatting for nested lists with code snippets
Apparently, there is a way to correctly format these,
just very careful formatting is needed for documents like this:
https://stackoverflow.com/questions/57366489/code-block-in-a-numbered-list-messes-up-numbering-python-markdown-mkdocs

The main trick is not to use ```, but use double-indent 4-spaces (8 total)
instead inside the desired list. This will render multi-line scripts with
proper `<pre>` tags instead of the inline code styling.