fix(processing): add integration test cases to validate that special files are not processed.
The squashfs filesystem contains the following special files:

- block device: not extracted because we're not running as root
- character device: not extracted because we're not running as root
- fifo device: extracted
- symlink: extracted
- socket: extracted

The fifo, symlink, and socket files are not processed by unblob,
otherwise it would hang (e.g. by waiting for data on the fifo).
Demonstrating that we have a proper fix.

Note that these special files cannot be tracked by git, so they're not
present in the output directory.