fix(handler/tar): replace 7zip with python tarfile module.
Given that 7zip just creates a text file with the symlink target path in
it, which is inaccurate, we had to switch to another extractor.

The decision was made to move to Python3's built-in tarfile module
given that's what we already use to calculate the chunk offsets.

tarfile module, like the tar and unar cli is unsafe when it comes to
symlink based traversal. This is no longer a problem since 97230c2 given
that unblob fixes link based traversal once the extraction has been
performed.