fix(squashfs): fix handling of SquashFS v4 big-endian image built by AVM.
AVM has a very interesting way of interpreting standards so we have to
use a different extractor. We identify whether an chunk is from AVM by
taking advantage of a tweak they do, documented at
https://raw.githubusercontent.com/Freetz/freetz/master/tools/make/squashfs4-host-be/AVM-BE-format.txt:

> super-block field 'mkfs_time' contains (instead of the filesystem
creation time) the same value as the super-block field 'bytes_used'

Depending on the AVM-ness, we use a different Command extractor to do
the job.