feat(elf) process carved ELF chunks in 2 steps
This is a report-only change affecting further processing of the report
file.

Previously carved out files were extracted in the same task, which
resulted in carved out ELF files not being recorded as separate "task"s.

With this change carved out ELF files will also appear inside "task"-s.

This is achieved in 2 steps, the first one creating a new file task:
- freshly carved ELF chunk is not extracted, but a new task
  is registered for the carved file, the carved file is always
  a whole file ELF chunk
- extract whole file ELF chunk

The above process is a change only for carved ELF chunks, which is
rather rare, normally ELF files are whole file chunks, which continue
to be extracted in one step.