fix(handlers): improve ZIP handler to support known indicators of ZIP64 usage
The implementation of the ZIP handler only checked one of the known
indicators of `ZIP64` usage. The missing logic resulted in the code
getting stuck in an infinite loop as it couldn't detect the end of the records
correctly.

Documentation for known indicators can be found at: https://pkware.cachefly.net/webdocs/APPNOTE/APPNOTE-6.3.1.TXT