fix tar handler: truncated tar files were not processed correctly
Truncated tar files had a performance impact due to an EOFError not
being caught and thus the tar file not being recognized, and this
got repeated for every file in the truncated tar file.

Details:

`tarfile.TarFile.next()` does not handle well, when `fileobj.seek`
raises an exception: it will not handle it, so the exception will get
escalated to the caller, which in turn did not handle it either causing
a truncated tar file skipped by the tar handler (in a rather expensive
way: throwing away significant work determining the tar chunk's end).

The below solution makes sure exceptions escaping tarfile.TarFile.next()
are handled.

----

`File.seek` was incorrectly converting `ValueError` to `EOFError`.
It is fixed by raising a `File.seek` specific subclass of `ValueError`.