chore(deps): replace ubi-reader fork by official PyPi version.
We cannot use git hosted dependencies if we want to publish unblob to
PyPi. We therefore moved from our fork of ubi-reader to the official
package of ubi-reader.

Our fork was synced with upstream except for one UI change we did to get
rid of superfluous directories created by ubireader_extract_images.
Since the maintainer of ubi-reader does not want to include it upstream,
we moved that logic into the UBIExtractor class.

See https://github.com/jrspruitt/ubi_reader/pull/67 for more details.