chore(deps): replace jefferson fork by the recently released PyPi package.
We're now the official maintainers of Jefferson and released the latest
version to PyPi. We therefore switched the dependency source from our
Github fork to PyPi.

Poetry identified a dependency mismatch between our required version of
cstruct (2.1) and the one required by Jefferson (5.2).

We upgraded cstruct to the latest version so that dependency resolution
works. The jump in major version may seem steep but cstruct maintainers
did not introduce API breaking changes.