feat(extractor): add in-house YAFFS/YAFFS2 extractor.
Initial objective was to get rid of yaffshiv as it was our last git
based dependency blocking us from publishing unblob to PyPi.

This change introduce an internal extractor for YAFFS and YAFFS2
formats, completely rewritten from scratch in order to provide proper
extraction capabilities.

The extractor relies on treelib to provide tree structures in Python.

That structure is used to load the filesystem representation in memory
and depth-first walks of it in order to re-create the filesystem on disk
during extraction.