fix [undeterministic] crash with specially crafted input
When the input is an archive of extraction with carved files,
the extraction paths overlap with the new extraction.
When a file was mmap-ed, and used in a worker, opening it again for
write in some other worker will truncate the file, which causes SIGBUS
on Linux in the worker mmap-ing the file.

This change prevents overwriting existing paths while carving chunks.