Process carved ELF chunks in two steps
Carved chunks are generally not important, so when they are fully
extracted, they can be deleted. ELF chunks are special, as they had
in general no extraction output. This is changed to have the extraction
output the ELF file itself by overriding chunk extraction.

Whole files have a single chunk that covers the file, they are also
processed differently from other chunks, as carving is skipped.

In the end ELF chunks are processed in two steps:
- carved files are simply copied, Handler defined extraction is not done
- non-carved whole files (copies of carved ELF chunks made previously
  are whole files now) are extracted as defined by the Handler