Make ChunkReport and UnknownChunkReport final
Newly introduced ElfChunk might set a precedent, and in the future
might get more of these special Chunk-s.

In the json based report output we just have a string representation
of the report class, but not the class hierarchy. It would become
non-trivial to filter for chunk reports, if variability would be let to
show up there as well.