Resolve initramfs start-end offset calculation ambiguity
Calculating the start-end offset is hard due to initramfs alignment
ambiguity, which might lead to 2 possible paddings.

This ambiguity could be resolved with extra code and knowledge of the
initramfs content (it is mostly a gzipped cpio archive).

Instead of including this extra complexity, the workaround is to carve
out at most 4 extra bytes. This 4 byte will appear as an extracted
unknown chunk.