fix(elf): handle non zero start_offset when calculating program/section end offset
Parsing the ELF section & program end offset only calculted the length, which
is actually the end offset when the ELF starts at the beginning of the file
hence start_offset is 0. In every other case we need to add the start_offset
to the calculated end to get the offset within the source file.

This only triggered a problem when the program or section was at the end
of the file, which is not always the case, usually the lookup tables are
located at the end.