fix(elf): use the carved out chunk in ramfs extraction
We should be using the carved out chunk as in case the ELF is not starting
on the 0 start offset of the source file, then the calculation will be off.

The easiest solution is to use the already carved out chunk file while
extracting the initramfs, so this way all the offset calculation will be
easier. Other option would be to pass the start_offset and use it in all
calculation or maybe use OffsetFile, but that currently does not support
offsetting the mmapped memory access.