feat(handler): Add support for HP IPKG format
The IPKG files are childs of a BDL file.

The header contains some informations such as name, version, revision, etc of the IPKG package.

After that header is a table of content that has the name, the offset, the size and a crc32 of each file it contains.

This repo https://github.com/tylerwhall/hpbdl contains information that was useful