feat(processing): Re-add non-POSIX file path handling
The sole reason of the existence and usage of the valid_path function
(to limit unblob not to handle files with non-POSIX path) was added
because of yara-python limitations in commit
e903602c94e4ca550911e3e68e01f0b09e48012e.

Since we are not using yara-python anymore, these limitations can now be
lifted!

The only consequence what I discovered during testing is that the
logger's format needed to be extended. It was already implemented to
encode Path-strings with surrogateescape, but paths can be logged in
strings as well (for example when logging the extractor command, which
contains the path of the file to be extracted).