fix(yaffs): improve data chunks storage and retrieval.
We made the wrong assumption that data chunks belonging to a header
chunk always come _after_ the header chunk. This is not true since a
truncated file could look like this:

- obsolete object header (chunk_id=0, obj_id=500, page_status=0)
- obsolete first chunk (chunk_id=1, obj_id=500, page_status=0)
- second chunk(chunk_id=2, obj_id=500)
- third chunk(chunk_id=3, obj_id=500)
- new object header (chunk_id=0, obj_id=500)
- new first chunk (chunk_id=1, obj_id=500)

To fix this, each chunk now possess an absolute offset value pointing to
where that chunk is located within the filesystem. We got rid of offsets
values within YAFFSEntry classes.

Instead of saving chunks within a YAFFSEntry object, we now use a
dedicated data_chunks attribute within the parser. It's a dictionnary
where keys are the object_id and the content is an array of chunks
belonging to that object_id.

With these changes we can handle chunks regardless of their location
towards their header chunk.