feat(processing): add multi-file directory handler
There are certain formats where the content is split between
multiple files. Currently unblob operates under the assumption
that all content resides in a single file.

A few examples where this might be relevant:
- multi-volume archives, such as 7zip, rar etc.
- VM snapshots
- content + index type formats

This change introduces a DirectoryHandler which can operate
on multiple files residing in one directory or at least under
one subtree. Most formats there is a "main" file which can
be identified by a directory file name pattern. Using this
first file the handler can identify the other files and
return a MultiFile object, similar to ValidChunks.

We do not support cases where a single file is part of
multiple MultiFile, also a file processed & extracted in
the context of a MultiFile is not processed by traditional
handlers. Also there is no carving step rather the files
are extracted directly into an extraction directory. The
original files are kept and never deleted, as these are normal
files, unlike carved out temporary chunks.

Files extracted from a MultiFile have a MultiFile as their
parent. This required extending the current File -> Chunk
reporting concept by introducing an abstract Blob
type which is the parent of Chunk and MultiFile as well.

MultiFileReports are reported under the directory Task,
but contains all included file paths as well.