feat(finder): hyperscan jump to the end of chunk
When we identify a chunk in a scanned file, we can safely ignore to the end
of the chunk to continue the pattern search there as an match or chunk
identified within the chunk would be anyhow discarded.
In case of some plaintext filesystem chunk, we can pick up many matches
so jumping to the end of the chunk would potentially save us from
many - at this point - irrelevant matches.