fix(extractor): stronger implementation of fix_extracted_directory.
1. recursive walking

We re-implemented fix_extracted_directory using os.listdir(). We do not
rely on os.walk because we need to fix the permissions recursively in
order to gain visibility into child directories. If we were to use
os.walk() on a directory containing directories on which we don't have
read permissions would mean we would not fix the content of those
directories.

2. handling ENAMETOOLONG

fix_extracted_directory now handles ENAMETOOLONG exception triggered on
symlinks with overly long target names. When that happens, we skip over
the file.

3. handling concurrent deletion

Someone (or something) can delete an object (file, directory, link, ...)
from the extraction directory as we are walking through it. We're also
handling such cases now so we don't raise FileNotFoundError.