Calculate and report weighted mean of entropy blocks
In most cases where entropy is calculated, and have more than 1K
input, the file is split up into multiple "blocks" of minimum 1K size,
and the entropy mean was calculated as the mean of the block-entropies.
This introduced a bias, as the last block is usually not the same size
as all the others, so its entropy had more say on the mean, than the
other blocks.
The above has more significant effect on files smaller than 80K, as that
is the limit, where a different block size can level the differences in
block sizes.

Let's look at an extreme example of an encrypted file of size 1025.
This would give us 2 "blocks" of sizes 1024 and 1 with entropies scaled
to "percentages" (0-100) for these blocks as ~100 and 0 respectively.

In this case naive mean is
  ~50 = (~100 + 0) / 2

in contrast, the weighted mean is a much better approximate:
  ~99.9 = (~100 * 1024 + 0 * 1) / (1024 + 1)