chore: use rust-overlay instead of fenix to get rid of IFD
Fenix uses a technique, called import from derivation, to acquire
details for a rust toolchain version: It looks-up the version number,
than downloads a manifest file, and build nix expressions from that
which will be imported at the end.

Nix flake check trips on IFD, with error messages like

> a 'aarch64-darwin' with features {} is required to build
'/nix/store/...', but I am a 'x86_64-linux' with features {}

Rust-overlay on the other hand ships with toolchain manifests, not
requiring IFD (but it is larger in size).

This problem is hidden at this moment, as the manifest from Rust 1.60
leaks into the nix store somehow on the builders.