fix(handler): handle false positives YAFFS sample by checking entries.
We have a YAFFS false positive sample that gets parsed but triggers a
NodeIDAbsentError when calling
'self.file_entries.expand_tree(mode=Tree.DEPTH)' during extraction. This
exception is raised by treelib if the tree is empty.

We handle this edge case by checking if the tree is empty after the
chunk identification, and raise an InvalidInputFormat if it is.