fix(handlers): handle tar archives with same file present more than once.
A "PermissionError: [Errno 13] Permission denied" is thrown when
expanding a tarfile to which a file had been added more than once, *if*
the file is not writeable.  (tarfile expands the file twice, but the
second time finds a non-writeable file in the way.)

Decided to handle this gracefully by catching the PermissionError and
logging it.

See https://bugs.python.org/issue30438 for a detailed discussion of this
issue.