fix(tar handler) extract duplicated entries
Tar files can be appended after creation, which also allows duplicate
entries/content for the same path.

A "PermissionError: [Errno 13] Permission denied" is thrown in TarFile
when expanding a tarfile to which a file had been added more than once,
*if* the file is not writeable.  (tarfile expands the file twice, but
the second time finds a non-writeable file in the way.)

This is supported from now on by removing the old file before
extracting the new file.

A remaining limitation is, that tar entry type change from directory to
non-directory will not work.
An error is raised and logged in this case, but the extraction will
continue.

----

See https://bugs.python.org/issue30438 for a detailed discussion of this
issue.