fix(handlers): fix infinite loop in zstd handler.
Given an invalid or truncated zstd file,
unblob/handlers/compression/zstd.py, in calculate_chunk(), did not check
for EOF in the while loop.

After reaching EOF, it always "reads" a block of type 0 and of size 0.
This is a valid block type (i.e. no raised exception) but not an "end
block" type, which is the only other condition to exit the loop.
Therefore, it kept looping actively and never returned.

Since we're reading from an mmap'ed file, reading from the file after
reaching EOF does not trigger an exception. Fixed by checking if the
read value is not empty and yielding an InvalidInputFormat exception
otherwise.

An integration test case reproducing the infinite loop has been added to
zstd test cases (truncated.dos.zstd).