fix(FileSystem): relax .problems attribute type
List[Report] and List[ExtractionProblem] are not compatible in general
due to mutability of the list could cause a non-ExtractionProblem Report
to appear in the FileSystem.problems.

However we do not expect to do anything with the FileSystem.problems
other than collecting them, so the specialized type is not really
important.
What we do lose is explicitly preventing the aliasing problem, whose
risk is considered very low.

In the end .problems can be passed in directly in ExtractResult,
without making an explicit copy in every case, just to get the type
checker happy.