fix(handlers): rely on pyfatfs to compute FAT images end offset
Some FAT samples that we have are truncated to match exactly the size of
reserved sectors + FAT sectors + active cluster sectors. It's like the
initial image minus the free clusters.

We switched our approach for end offset calculation by simply
calculating the offset of the highest cluster within the filesystem. The
end offset corresponds to the offset of that cluster + the size of the
file located there.

Parsing is performed by pyfatfs, which implements similar structure
validation so whenever an error raised we yield an InvalidInputFormat.