fix(handlers): rely on pyfatfs to compute FAT images end offset
Some FAT samples that we have are truncated to match exactly the size of
reserved sectors + FAT sectors + active cluster sectors. It's like the
initial image minus the free clusters.

We switched our approach for end offset calculation by simply relying on
the File Allocation Table. Starting from the end of the FAT, we find the
last non-freed (different than 0) cluster index. This cluster is the
last cluster with data within the filesystem, we then translate the
cluster index to an offset, that corresponds to the end offset.

Parsing is performed by pyfatfs, which implements similar structure
validation so whenever an error raised we yield an InvalidInputFormat.