fix(handlers): add support for unix-compatible (aka v7) tar files.
v7 tar headers do not have the 'ustar' magic that we match on for modern
tar files. In order to match on those v7 archive, we build a regular
expression that matches on mode, uid, gid, mtime, size files given their
properties:

- fixed size (e.g. 8 for mode)
- optionally prepended by whitespaces
- suffixed by null bytes (null terminated)
- ASCII encoded octal digits (0x30 to 0x37)

In order to build a pattern that can be handled by hyperscan without
using a notation such as '[\w]{1,99}' for path name (see below for
detailed explanation), we rely on utility function to build a regular
expression to match all possible combination using the or (|) operator.

Note: hyperscan will yield a "Pattern is too large" exception when
trying to use '{1,99}' notation. Even though we found out that using
'.*' works, it would have an important performance impact on pattern
matching. That's why we decided to go with the OR operator approach with
combination.

See https://intel.github.io/hyperscan/dev-reference/compilation.html for
more information about this.