ci: do not build and push Docker images on pull requests
Build the docker images only when pushing to main, there's no need to
build those images in each and every pull request.

Initial trigger for this was that dependabot does not possess the
required privileges to push to our Docker registry, so there's no point
building and pushing the image at that point.