fix(handler): improve zip64 detection
We improve the zip64 detection by trying to parse the zip64 early.
If the zip file contains a zip64 header, we are sure its a zip64.
If it doesn't hold a zip64 header, we are sure its not a zip64 and we fall back to zip32.

In order to create the test files, we used this command
```
cat somefile.txt | zip > zip64.zip
zip -F zip64.zip --out zip64-without-cd.zip
```