fix(yaffs): extend potential page size list to include 2032 bytes
In a sample we see a page_size with 2032 bytes and 16 bytes spare. Apparently
the sum of page+spare is 2048 bytes.

As we have many checks to see if a page_size is indeed valid and calculate
also spare size, having an additional potential page size should not cause
any problem.

As page sizes are tested in the order specified, this new and unlikely case
added to the end not to slow down other extractions.