nix: use `devenv` to create a shell environment
Devenv[^1] is tool to create easily composable development
environments for nix.

Hooking it in to our nix flake gives us the following improvements:
 - more declarative python environment config
 - avoid the `patchElf` hack to pull in native dependencies
 - python packages updates won't undo the effect of `patchElf`
 - due to a long-standing bug[^2], a patched `devenv` version is needed

The `.devenv` directory is pre-created because of a quirk: it needs to
be present in the `src` derivation of `shell.nix`, otherwise it is
unable to create it due to a permission error.

[^1]: https://devenv.sh/
[^2]: https://github.com/cachix/devenv-nixpkgs/issues/2