chore: eliminate _FileTask.carve_dir
`_FileTask.carve_dir` was initially used for both extraction and carving.
The naming of the directories can now differ, so it is not used anymore
apart from an existence check, which would terminate this branch of the
extraction. This output directory existence check is now present in both
the carving and extraction paths, and the output report's name is also
renamed, to accommodate both types of output directories.

`ExtractDirectoryExistsReport` was generalized to
`OutputDirectoryExistsReport` instead of introducing yet another
`Report` type - `CarveDirectoryExistsReport`.