chore: mv unblob python/unblob
This is done to circumvent an import resolution ambiguity[^1]:
When executing Python scripts such as `pytest` or `unblob` itself, the
current directory will be placed into `sys.path`, so imports from
`unblob` packages short-circuit and will be resolved from the source
root, and won't go through the virtualenv at all.

This makes it impossible for imports to be resolved from the
virtualenv, even if a non-editable unblob package is installed to
it. This is less of an issue in most cases, as the package mostly a
1-to-1 copy of the source root. However, with extension modules,
the following test setup just won't work:

- build wheel
- create virtualenv
- install wheel
- run pytest inside the repo root

We are using this flow for the docker image and Nix builds as well.

[^1]: https://www.maturin.rs/project_layout.html#mixed-rustpython-project