fix(report): adapt landlock passthrough for report file
Report writing only worked if the report was being written to a child
location of the extract directory. This was a byproduct of setting
rw permissions on the extraction directory.

It did not work when the report file was being written to a location
that is unrelated to the extraction directory.

Fixed by requesting rw access to the report's file parent directory.

I tried to get smart by only allowing make_reg and read_write on the
file, but it never fully worked because of file truncation and the
fact that LANDLOCK_ACCESS_FS_TRUNCATE is only available since ABI
version 3 in landlock.